fix incorrect class fix rrd rename logic
