+++ LibreNMS/OS/Coriant.php
@@ -26,8 +26,8 @@
 namespace LibreNMS\OS;
 
 use App\Models\TnmsneInfo;
+use App\Observers\ModuleModelObserver;
 use LibreNMS\Interfaces\Polling\OSPolling;
-use LibreNMS\Util\ModuleModelObserver;
 use Log;
 
 class Coriant extends \LibreNMS\OS implements OSPolling
