+++ app/Observers/ModuleModelObserver.php
@@ -1,6 +1,6 @@
 <?php
-/**
- * DiscoveryModelObserver.php
+/*
+ * ModuleModelObserver.php
  *
  * Displays +,-,U,. while running discovery and adding,deleting,updating, and doing nothing.
  *
@@ -18,11 +18,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @link       http://librenms.org
- * @copyright  2018 Tony Murray
+ * @copyright  2020 Tony Murray
  * @author     Tony Murray <murraytony@gmail.com>
  */
 
-namespace LibreNMS\Util;
+namespace App\Observers;
 
 use Illuminate\Database\Eloquent\Model as Eloquent;
