+++ tests/data/aix.json
@@ -533,7 +533,7 @@
                     "mempool_index": "10",
                     "entPhysicalIndex": null,
                     "mempool_type": "hrstorage",
-                    "mempool_class": "virtual",
+                    "mempool_class": "swap",
                     "mempool_precision": 4096,
                     "mempool_descr": "/dev/hd6",
                     "mempool_perc": 4,
