+++ tests/data/aix_net-snmp.json
@@ -155,7 +155,7 @@
                     "mempool_index": "10",
                     "entPhysicalIndex": null,
                     "mempool_type": "hrstorage",
-                    "mempool_class": "virtual",
+                    "mempool_class": "swap",
                     "mempool_precision": 4096,
                     "mempool_descr": "Swap space",
                     "mempool_perc": 0,
