+++ tests/data/aix_std.json
@@ -961,7 +961,7 @@
                     "mempool_index": "21",
                     "entPhysicalIndex": null,
                     "mempool_type": "hrstorage",
-                    "mempool_class": "virtual",
+                    "mempool_class": "swap",
                     "mempool_precision": 4096,
                     "mempool_descr": "/dev/hd6",
                     "mempool_perc": 0,
