+++ tests/data/dsm_ds214se.json
@@ -814,7 +814,7 @@
                     "mempool_index": "10",
                     "entPhysicalIndex": null,
                     "mempool_type": "hrstorage",
-                    "mempool_class": "virtual",
+                    "mempool_class": "swap",
                     "mempool_precision": 1024,
                     "mempool_descr": "Swap space",
                     "mempool_perc": 0,
