+++ tests/data/solidserver.json
@@ -1233,7 +1233,7 @@
                     "mempool_index": "10",
                     "entPhysicalIndex": null,
                     "mempool_type": "hrstorage",
-                    "mempool_class": "virtual",
+                    "mempool_class": "swap",
                     "mempool_precision": 4096,
                     "mempool_descr": "Swap space (total)",
                     "mempool_perc": 1,
